Develop

+++ src/components/Admin/UsersTable.js
@@ -2,8 +2,10 @@ import React, { useState } from 'react';
 import Table from 'react-bootstrap/Table';
 import Button from 'react-bootstrap/Button';
 import FormCheck from 'react-bootstrap/FormCheck';
+import styled from 'styled-components';
 import { LoadableButton } from '../Loadable';
 import Router from 'next/router';
+import Link from 'next/link';
 
 const UserRow = ({ key, user }) => {
   const [email, setEmail] = useState(user.email);
@@ -45,8 +47,12 @@ const UserRow = ({ key, user }) => {
 
   const handleDelete = async (event) => {
     event.preventDefault();
-    setLoading(true);
+    const msg = `¿Seguro que querés borrar el usuario ${user.email}?`;
+    if (!confirm(msg)) {
+      return;
+    }
 
+    setLoading(true);
     const res = await fetch(`/api/users/${user._id}`, {
       method: 'DELETE',
       headers: {
@@ -138,53 +144,79 @@ const UserRow = ({ key, user }) => {
       </td>
       <td>
         {modified ? (
-          <LoadableButton
-            loading={loading}
-            variant="success"
-            onClick={handleSave}
-          >
-            Guardar
-          </LoadableButton>
+          <>
+            <LoadableButton
+              disabled={loading}
+              loading={loading}
+              variant="success"
+              onClick={handleSave}
+            >
+              Guardar
+            </LoadableButton>
+            <Button
+              disabled={loading}
+              variant="secondary"
+              onClick={handleCancel}
+            >
+              Cancelar
+            </Button>
+          </>
         ) : (
-          <LoadableButton
-            loading={loading}
-            variant="danger"
-            onClick={handleDelete}
-          >
-            Eliminar
-          </LoadableButton>
+          <>
+            <LoadableButton
+              disabled={loading}
+              loading={loading}
+              variant="danger"
+              onClick={handleDelete}
+            >
+              Eliminar
+            </LoadableButton>
+            <Link href={`/editar-perfil/${user.username}`}>
+              <Button disabled={loading} variant="primary">
+                Editar
+              </Button>
+            </Link>
+          </>
         )}
-        <Button
-          loading={loading}
-          variant="secondary"
-          onClick={handleCancel}
-          style={{ visibility: modified ? 'visible' : 'hidden' }}
-        >
-          Cancelar
-        </Button>
       </td>
     </tr>
   );
 };
 
-const UsersTable = ({ users }) => (
-  <Table striped bordered hover>
-    <thead>
-      <tr>
-        <th>#</th>
-        <th>Mail</th>
-        <th>Password</th>
-        <th>Rol</th>
-        <th>Permisos</th>
-        <th>Acciones</th>
-      </tr>
-    </thead>
-    <tbody>
-      {users.map((user) => (
-        <UserRow key={user._id} user={user} />
-      ))}
-    </tbody>
-  </Table>
-);
+const FloatingButton = styled(LoadableButton)`
+  position: absolute;
+  right: 0;
+  top: -4em;
+  padding: 1em;
+`;
+
+const UsersTable = ({ users }) => {
+  return (
+    <Table striped bordered hover>
+      <FloatingButton
+        href="/crear-perfil"
+        variant="success"
+        style={{ position: 'absolute' }}
+      >
+        Crear Nuevo Perfil
+      </FloatingButton>
+      <thead>
+        <tr>
+          <th>#</th>
+          <th>Mail</th>
+          <th>Password</th>
+          <th>Rol</th>
+          <th>Permisos</th>
+          <th>Acciones</th>
+        </tr>
+      </thead>
+      <tbody>
+        {users.map((user) => (
+          <UserRow key={user._id} user={user} />
+        ))}
+      </tbody>
+    </Table>
+  );
+};
 
 export default UsersTable;
